+++ Functions/Environment.ps1
@@ -15,9 +15,9 @@ function GetPesterOs
     {
         'Windows'
     }
-    elseif (Get-Variable -Name 'IsOSX' -ErrorAction 'SilentlyContinue' -ValueOnly )
+    elseif (Get-Variable -Name 'IsMacOS' -ErrorAction 'SilentlyContinue' -ValueOnly )
     {
-        'OSX'
+        'macOS'
     }
     elseif (Get-Variable -Name 'IsLinux' -ErrorAction 'SilentlyContinue' -ValueOnly )
     {